requirements.txt
debugpy 
numpy 
matplotlib 
scikit-image 
albumentations 
tqdm 
pandas 
torchmetrics 
tensorboard 
imagecodecs
transformers[torch]
pillow
pycocotools

timm
accelerate

opencv-python==4.8.0.74
onnxruntime
faster-coco-eval

datasets
wandb
scikit-learn

lightning
gdown